Improve badge page
